contracts/requirements.txt
pyteal==0.26.1
algosdk==2.6.1
python-dotenv==1.0.1
